move period slippage to src/fetch
